canbus: fix compile warning.

+++ modules/canbus/vehicle/lexus/protocol/accel_cmd_100.cc
@@ -61,7 +61,7 @@ Accelcmd100* Accelcmd100::set_ignore_overrides(bool ignore_overrides) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 1,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Accelcmd100::set_p_ignore_overrides(uint8_t* data, bool ignore_overrides) {
-  int x = ignore_overrides;
+  uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 1, 1);
@@ -76,7 +76,7 @@ Accelcmd100* Accelcmd100::set_enable(bool enable) {
 // 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 0, 'type': 'bool',
 // 'order': 'motorola', 'physical_unit': ''}
 void Accelcmd100::set_p_enable(uint8_t* data, bool enable) {
-  int x = enable;
+  uint8_t x = enable;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 0, 1);
@@ -91,7 +91,7 @@ Accelcmd100* Accelcmd100::set_clear_override(bool clear_override) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 2,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Accelcmd100::set_p_clear_override(uint8_t* data, bool clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 2, 1);
@@ -106,7 +106,7 @@ Accelcmd100* Accelcmd100::set_clear_faults(bool clear_faults) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 3,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Accelcmd100::set_p_clear_faults(uint8_t* data, bool clear_faults) {
-  int x = clear_faults;
+  uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
   to_set.set_value(static_cast<uint8_t>(x), 3, 1);
