+++ modules/canbus/vehicle/lexus/protocol/cruise_control_buttons_cmd_108.cc
@@ -74,7 +74,7 @@ Cruisecontrolbuttonscmd108::set_cruise_control_button(
 void Cruisecontrolbuttonscmd108::set_p_cruise_control_button(
     uint8_t* data, Cruise_control_buttons_cmd_108::Cruise_control_buttonType
                        cruise_control_button) {
-  int x = cruise_control_button;
+  uint8_t x = cruise_control_button;
 
   Byte to_set(data + 1);
   to_set.set_value(x, 0, 8);
@@ -91,7 +91,7 @@ Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_ignore_overrides(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Cruisecontrolbuttonscmd108::set_p_ignore_overrides(uint8_t* data,
                                                         bool ignore_overrides) {
-  int x = ignore_overrides;
+  uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 1, 1);
@@ -108,7 +108,7 @@ Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_clear_override(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Cruisecontrolbuttonscmd108::set_p_clear_override(uint8_t* data,
                                                       bool clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 2, 1);
@@ -124,7 +124,7 @@ Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_enable(
 // 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 0, 'type': 'bool',
 // 'order': 'motorola', 'physical_unit': ''}
 void Cruisecontrolbuttonscmd108::set_p_enable(uint8_t* data, bool enable) {
-  int x = enable;
+  uint8_t x = enable;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 0, 1);
@@ -141,7 +141,7 @@ Cruisecontrolbuttonscmd108* Cruisecontrolbuttonscmd108::set_clear_faults(
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Cruisecontrolbuttonscmd108::set_p_clear_faults(uint8_t* data,
                                                     bool clear_faults) {
-  int x = clear_faults;
+  uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 3, 1);
